no longer delete domain doc

+++ src/routes/api/admin/users/delete.js
@@ -12,7 +12,8 @@ router.get('/api/admin/delete/:id', passwordAuthentication, async (req, res) => 
     if (id) {
       const userData = await req.app.server.models.UserModel.findOne({ id: id });
       if (userData) {
-        if(userData !== req.app.server.defaults.domain) {
+        const domainData = await req.app.server.models.DomainModel.findOne({ domain: userData.domain.domain });
+        if(userData.domain.domain !== req.app.server.defaults.domain && domainData.owner == userData.id) {
           await req.app.server.models.DomainModel.deleteOne({ domain: userData.domain.domain });
         }
         await req.app.server.models.UserModel.deleteOne({ id: id });
