+++ src/routes/api/user/account.js
@@ -13,7 +13,8 @@ router.post('/api/user/account', async (req, res) => {
   if(string === stringCheck) {
     bcrypt.compare(password, req.session.userData.authentication.password).then(async e => {
       if(e) {
-        if(req.session.userData.domain.domain !== req.app.server.defaults.domain) {
+        const domainData = await req.app.server.models.DomainModel.findOne({ domain: req.session.userData.domain.domain });
+        if(req.session.userData.domain.domain !== req.app.server.defaults.domain && domainData.owner == req.session.userData.id) {
           await req.app.server.models.DomainModel.deleteOne({ domain: req.session.userData.domain.domain });
         }
         await req.app.server.models.UserModel.deleteOne({ id: req.session.userData.id });
